adding payment repository

+++ server/ShoppingSystem/src/main/java/lt/mif/ise/domain/Product.java
@@ -27,7 +27,7 @@ public class Product {
     private String imageUrl;
     
     @Column(nullable= false, precision=8, scale=3)
-    private BigDecimal price;
+    private int price;
 
     public void setName(String name) {
         this.name = name;
@@ -45,11 +45,11 @@ public class Product {
         this.id = id;
     }
 
-	public BigDecimal getPrice() {
+	public int getPrice() {
 		return price;
 	}
 
-	public void setPrice(BigDecimal price) {
+	public void setPrice(int price) {
 		this.price = price;
 	}
